requirements.txt
pygame==2.5.2
pygame-menu==4.4.3
pillow==10.2.0